backend/requirements.txt
openai
langchain-openai
langchain
pinecone-client
fastapi
uvicorn
starlette
python-dotenv